+++ tests/models/test_modeling_common.py
@@ -858,11 +858,6 @@ class ModelTesterMixin:
     ):
         if not self.model_class._supports_gradient_checkpointing:
             return  # Skip test if model does not support gradient checkpointing
-        if self.model_class.__name__ in [
-            "UNetSpatioTemporalConditionModel",
-            "AutoencoderKLTemporalDecoder",
-        ]:
-            return
 
         init_dict, inputs_dict = self.prepare_init_args_and_inputs_for_common()
