+++ tests/pipelines/test_pipelines_common.py
@@ -48,7 +48,7 @@ from diffusers.utils.testing_utils import (
     torch_device,
 )
 
-from ..models.autoencoders.test_models_vae import (
+from ..models.autoencoders.vae import (
     get_asym_autoencoder_kl_config,
     get_autoencoder_kl_config,
     get_autoencoder_tiny_config,
